+++ src/compare.rs
@@ -17,13 +17,12 @@ use crate::{
     flex_enums::{cash_action, CashAction},
     flex_query::{CashTransaction, FlexQueryResponse},
     flex_reader::load_report,
-    ledger_runner,
+    ledger_runner::{self, get_ledger_start_date},
     model::CommonTransaction,
     ISO_DATE_FORMAT,
 };
 
 pub const TRANSACTION_DAYS: u8 = 60;
-//pub(crate) const DATE_MODE: &str = "book"; // "book" / "effective"
 
 /**
  * Compares transactions in the downloaded IB Flex report to Ledger.
@@ -41,12 +40,12 @@ pub fn compare(params: CompareParams) -> anyhow::Result<String> {
         return Ok(msg.into());
     }
 
+    // identify the start date for the tx range:
+    let start_date = get_oldest_ib_date(&ib_txs, params.effective_dates);
+
     // get_ledger_tx
-    let ledger_txs = ledger_runner::get_ledger_tx(
-        cfg.ledger_init_file,
-        params.comparison_date,
-        params.effective_dates,
-    );
+    let ledger_txs =
+        ledger_runner::get_ledger_tx(cfg.ledger_init_file, start_date, params.effective_dates);
     log::debug!("Found {} Ledger transactions", ledger_txs.len());
 
     // compare
@@ -58,19 +57,14 @@ pub fn compare(params: CompareParams) -> anyhow::Result<String> {
 fn compare_txs(
     ib_txs: Vec<CommonTransaction>,
     ledger_txs: Vec<CommonTransaction>,
-    use_effective_dates: bool,
+    use_effective_date: bool,
 ) -> anyhow::Result<String> {
     let mut result = String::default();
 
     for ibtx in ib_txs {
         log::debug!("Matching ib tx: {:?}", ibtx);
 
-        let ib_comparison_date = if use_effective_dates {
-            ibtx.date.format(ISO_DATE_FORMAT).to_string()
-        } else {
-            // actual date
-            ibtx.report_date.to_owned()
-        };
+        let ib_comparison_date = get_comparison_date(&ibtx, use_effective_date);
         log::debug!("using ib date: {:?}", ib_comparison_date);
 
         let matches: Vec<&CommonTransaction> = ledger_txs
@@ -96,6 +90,31 @@ fn compare_txs(
     println!("Complete.");
 
     Ok(result)
+}
+
+fn get_comparison_date(ibtx: &CommonTransaction, use_effective_date: bool) -> String {
+    match use_effective_date {
+        true => ibtx.date.format(ISO_DATE_FORMAT).to_string(),
+        false => ibtx.report_date.to_owned(), // actual date
+    }
+}
+
+/// Finds the date of the oldest transaction in the report.
+/// This date is to be used for time-boxing Ledger report.
+fn get_oldest_ib_date(ib_txs: &Vec<CommonTransaction>, use_effective_date: bool) -> String {
+    if ib_txs.is_empty() {
+        return get_ledger_start_date(None);
+    }
+
+    //ib_txs.sort_unstable_by_key(|ibtx| get_comparison_date(&ibtx, use_effective_date) );
+    let oldest_record = ib_txs
+        .iter()
+        .min_by_key(|ibtx| get_comparison_date(&ibtx, use_effective_date))
+        .expect("got oldest date");
+
+    log::debug!("oldest tx: {:?}", oldest_record);
+
+    get_comparison_date(&oldest_record, use_effective_date)
 }
 
 /// Load the symbol mappings.
@@ -216,7 +235,6 @@ fn read_flex_report(cfg: &Config) -> Vec<CashTransaction> {
  */
 #[derive(Debug)]
 pub struct CompareParams {
-    pub comparison_date: Option<String>,
     pub flex_report_path: Option<String>,
     pub flex_reports_dir: Option<String>,
     pub ledger_init_file: Option<String>,
@@ -230,12 +248,10 @@ pub struct CompareParams {
 mod tests {
     use std::path::PathBuf;
 
-    use super::{compare, compare_txs, load_symbols};
+    use super::{compare, load_symbols};
     use crate::{
-        compare::{convert_ib_txs, get_ib_tx, CompareParams},
-        config::Config,
+        compare::{convert_ib_txs, CompareParams},
         flex_query::CashTransaction,
-        ledger_runner::get_ledger_tx,
         test_fixtures::*,
     };
 
@@ -259,20 +275,19 @@ mod tests {
         assert!(!ib_tx.is_empty());
     }
 
-    #[rstest::rstest]
-    #[test_log::test]
-    fn test_compare_tx(cmp_config: Config) {
-        let ib_txs = get_ib_tx(&cmp_config);
-        let use_effective_dates = false;
-        let ledger_txs = get_ledger_tx(cmp_config.ledger_init_file, None, use_effective_dates);
+    // #[rstest::rstest]
+    // #[test_log::test]
+    // fn test_compare_tx(cmp_config: Config) {
+    //     let ib_txs = get_ib_tx(&cmp_config);
+    //     let use_effective_dates = false;
+    //     let ledger_txs = get_ledger_tx(cmp_config.ledger_init_file, None, use_effective_dates);
 
-        log::debug!("comparing {:?} *** and *** {:?}", ib_txs, ledger_txs);
+    //     log::debug!("comparing {:?} *** and *** {:?}", ib_txs, ledger_txs);
 
-        let actual = compare_txs(ib_txs, ledger_txs, use_effective_dates);
+    //     let actual = compare_txs(ib_txs, ledger_txs, use_effective_dates);
 
-        assert!(actual.is_ok());
-        // assert!(false)
-    }
+    //     assert!(actual.is_ok());
+    // }
 
     #[rstest::rstest]
     #[test_log::test]
@@ -292,7 +307,6 @@ mod tests {
     #[test_log::test]
     fn test_compare_w_multiple_matches() {
         let cmp_params = CompareParams {
-            comparison_date: None,
             flex_report_path: Some("tests/tax_adj_report.xml".into()),
             flex_reports_dir: None,
             ledger_init_file: Some("tests/tax_adj.ledgerrc".into()),
@@ -312,7 +326,6 @@ mod tests {
     #[test_log::test]
     fn test_compare_w_multiple_matches_effective_dates() {
         let cmp_params = CompareParams {
-            comparison_date: None,
             flex_report_path: Some("tests/tax_adj_report.xml".into()),
             flex_reports_dir: None,
             ledger_init_file: Some("tests/tax_adj.ledgerrc".into()),