+++ src/main.rs
@@ -29,7 +29,6 @@ async fn main() {
 
         Commands::Cmp(params) => {
             let cmp_params = CompareParams {
-                comparison_date: params.comparison_date.to_owned(),
                 flex_report_path: params.flex_report_path.to_owned(),
                 flex_reports_dir: params.flex_reports_dir.to_owned(),
                 ledger_init_file: params.ledger_init_file.to_owned(),